Designate feature mixin classes & employ class_prepared signal to register features
